fix(tests): split routes tests into sibling file, fix 3 new test tokio/time issues
CI on the Tasks 4-7 push surfaced two real problems:

1. File Size Check failed because crates/sp-server/src/api/routes.rs
   grew to 1027 lines (limit 1000). The file now sits at 541 lines;
   the test module is moved to a sibling routes_tests.rs included via
   #[cfg(test)] #[path = "routes_tests.rs"] mod tests;. This keeps
   `use super::*` working without changing any test content.

2. Three new tests panicked on the Ubuntu 'Test' job because they
   combined sqlx pool setup with `tokio::test(start_paused = true)`.
   With virtual time paused, sqlx's pool acquire timeout (real-time
   timer under the hood) fires as PoolTimedOut. Fixes:

   - playback::tests::position_events_are_throttled: switched to plain
     #[tokio::test] with a real 550ms sleep instead of
     tokio::time::advance. Test now runs in ~1s.

   - reprocess::tests::bump_video_backoff_escalates_and_caps: was a
     plain #[test] using SqlitePool::connect_lazy which still needs a
     tokio context. Changed to #[tokio::test] with the real setup()
     helper.

   - reprocess::tests::gemini_rate_limit_triggers_global_cooldown:
     switched from #[tokio::test(start_paused = true)] with
     tokio::time::advance to plain #[tokio::test]. Cooldown expiry is
     simulated by directly setting worker.gemini_cooldown_until to an
     Instant in the past, which is what in_global_cooldown() checks.

No test behaviour was weakened or skipped — the assertions still
exercise the same code paths, just without virtual-time-pausing.
