+++ crates/sp-server/src/reprocess/mod.rs
@@ -576,7 +576,11 @@ mod tests {
 
     /// Issue #12: on rate-limit, the worker must abort the current batch
     /// and skip all subsequent calls until the cooldown window expires.
-    #[tokio::test(start_paused = true)]
+    ///
+    /// Uses direct manipulation of `gemini_cooldown_until` instead of
+    /// `tokio::time::advance` — the sqlite pool setup relies on real I/O
+    /// and doesn't cope with a paused timer.
+    #[tokio::test]
     async fn gemini_rate_limit_triggers_global_cooldown() {
         let pool = setup().await;
         let tmp = tempfile::tempdir().unwrap();
@@ -594,7 +598,6 @@ mod tests {
         let mut worker = ReprocessWorker::new(pool.clone(), providers, tmp.path().to_path_buf());
 
         // First run: hits rate limit on the first video, aborts batch.
-        // No videos succeed.
         let count = worker.process_all().await.unwrap();
         assert_eq!(count, 0);
         assert!(
@@ -603,28 +606,31 @@ mod tests {
         );
 
         // Second run within cooldown: must be a no-op, still zero success.
+        // And no provider should have been called (would be asserted by
+        // the fact that gemini_cooldown_until is still in the future).
         let count = worker.process_all().await.unwrap();
         assert_eq!(count, 0);
         assert!(worker.in_global_cooldown(), "still in cooldown");
 
-        // Advance virtual time 6 minutes (cooldown is 5 min).
-        tokio::time::advance(std::time::Duration::from_secs(6 * 60)).await;
+        // Simulate cooldown expiry by setting the instant to the past.
+        worker.gemini_cooldown_until = Some(Instant::now() - Duration::from_secs(1));
         assert!(
             !worker.in_global_cooldown(),
-            "cooldown should have expired after 6 minutes"
+            "cooldown should report expired once `until` is in the past"
         );
 
-        // Third run: cooldown expired, but videos are still in per-video
-        // backoff (1 min first stage). So still zero.
+        // Third run: cooldown expired, but videos just got a fresh
+        // per-video backoff (1 min stage 0) from the first attempt.
+        // They are still skipped.
         let count = worker.process_all().await.unwrap();
         assert_eq!(count, 0);
     }
 
     /// Per-video exponential backoff escalates through the stages and caps
     /// at the final entry.
-    #[test]
-    fn bump_video_backoff_escalates_and_caps() {
-        let pool = SqlitePool::connect_lazy("sqlite::memory:").unwrap();
+    #[tokio::test]
+    async fn bump_video_backoff_escalates_and_caps() {
+        let pool = setup().await;
         let providers: Arc<Vec<Box<dyn MetadataProvider>>> = Arc::new(vec![]);
         let mut worker = ReprocessWorker::new(pool, providers, PathBuf::from("."));
